cacaoq/requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
pdfplumber>=0.11.0
yfinance>=0.2.36
httpx>=0.27.0
pandas>=2.1.0
numpy>=1.26.0
watchdog>=4.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
